backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
PyMuPDF==1.24.9
google-generativeai==0.3.0
scikit-learn==1.4.0
python-multipart==0.0.6
python-dotenv==1.0.0
reportlab==4.0.0
numpy>=1.24.0